refactor: ViewModelProvider의 factory를 ViewModel의 동반객체로 이동
