--- android/app/src/main/java/com/now/naaga/presentation/adventureresult/AdventureResultFactory.kt
@@ -1,20 +1,0 @@
-package com.now.naaga.presentation.adventureresult
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.now.domain.repository.AdventureRepository
-import com.now.domain.repository.RankRepository
-
-class AdventureResultFactory(
-    private val adventureRepository: AdventureRepository,
-    private val rankRepository: RankRepository,
-) :
-    ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(AdventureResultViewModel::class.java)) {
-            return AdventureResultViewModel(adventureRepository, rankRepository) as T
-        } else {
-            throw IllegalArgumentException()
-        }
-    }
-}
