+++ android/app/src/main/java/com/now/naaga/presentation/adventureresult/AdventureResultViewModel.kt
@@ -3,9 +3,12 @@ package com.now.naaga.presentation.adventureresult
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import com.now.domain.model.AdventureResult
 import com.now.domain.repository.AdventureRepository
 import com.now.domain.repository.RankRepository
+import com.now.naaga.data.repository.DefaultAdventureRepository
+import com.now.naaga.data.repository.DefaultRankRepository
 import com.now.naaga.data.throwable.DataThrowable
 import com.now.naaga.data.throwable.DataThrowable.GameThrowable
 
@@ -48,6 +51,20 @@ class AdventureResultViewModel(
         when (throwable) {
             is GameThrowable -> { _errorMessage.value = throwable.message }
             else -> {}
+        }
+    }
+
+    companion object {
+        val Factory = AdventureResultFactory(DefaultAdventureRepository(), DefaultRankRepository())
+
+        class AdventureResultFactory(
+            private val adventureRepository: AdventureRepository,
+            private val rankRepository: RankRepository,
+        ) :
+            ViewModelProvider.Factory {
+            override fun <T : ViewModel> create(modelClass: Class<T>): T {
+                return AdventureResultViewModel(adventureRepository, rankRepository) as T
+            }
         }
     }
 }
