+++ android/app/src/main/java/com/now/naaga/presentation/mypage/MyPageActivity.kt
@@ -9,9 +9,6 @@ import androidx.lifecycle.ViewModelProvider
 import com.now.naaga.data.firebase.analytics.AnalyticsDelegate
 import com.now.naaga.data.firebase.analytics.DefaultAnalyticsDelegate
 import com.now.naaga.data.firebase.analytics.MYPAGE_GO_RESULTS
-import com.now.naaga.data.repository.DefaultPlaceRepository
-import com.now.naaga.data.repository.DefaultRankRepository
-import com.now.naaga.data.repository.DefaultStatisticsRepository
 import com.now.naaga.databinding.ActivityMyPageBinding
 import com.now.naaga.presentation.adventurehistory.AdventureHistoryActivity
 
@@ -31,11 +28,7 @@ class MyPageActivity : AppCompatActivity(), AnalyticsDelegate by DefaultAnalytic
     }
 
     private fun initViewModel() {
-        val rankRepository = DefaultRankRepository()
-        val statisticsRepository = DefaultStatisticsRepository()
-        val placeRepository = DefaultPlaceRepository()
-        val factory = MyPageFactory(rankRepository, statisticsRepository, placeRepository)
-        viewModel = ViewModelProvider(this, factory)[MyPageViewModel::class.java]
+        viewModel = ViewModelProvider(this, MyPageViewModel.Factory)[MyPageViewModel::class.java]
         binding.viewModel = viewModel
         binding.lifecycleOwner = this
     }
