--- android/app/src/main/java/com/now/naaga/presentation/mypage/MyPageFactory.kt
@@ -1,21 +1,0 @@
-package com.now.naaga.presentation.mypage
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.now.domain.repository.PlaceRepository
-import com.now.domain.repository.RankRepository
-import com.now.domain.repository.StatisticsRepository
-
-class MyPageFactory(
-    private val rankRepository: RankRepository,
-    private val statisticsRepository: StatisticsRepository,
-    private val placeRepository: PlaceRepository,
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(MyPageViewModel::class.java)) {
-            return MyPageViewModel(rankRepository, statisticsRepository, placeRepository) as T
-        } else {
-            throw IllegalArgumentException()
-        }
-    }
-}
