--- android/app/src/main/java/com/now/naaga/presentation/rank/RankFactory.kt
@@ -1,16 +1,0 @@
-package com.now.naaga.presentation.rank
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.now.domain.repository.RankRepository
-
-class RankFactory(private val rankRepository: RankRepository) :
-    ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(RankViewModel::class.java)) {
-            return RankViewModel(rankRepository) as T
-        } else {
-            throw IllegalArgumentException()
-        }
-    }
-}
