+++ android/app/src/main/java/com/now/naaga/presentation/rank/RankViewModel.kt
@@ -3,10 +3,12 @@ package com.now.naaga.presentation.rank
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import com.now.domain.model.OrderType
 import com.now.domain.model.Rank
 import com.now.domain.model.SortType
 import com.now.domain.repository.RankRepository
+import com.now.naaga.data.repository.DefaultRankRepository
 import com.now.naaga.data.throwable.DataThrowable
 import com.now.naaga.data.throwable.DataThrowable.PlayerThrowable
 
@@ -56,6 +58,17 @@ class RankViewModel(private val rankRepository: RankRepository) : ViewModel() {
         when (throwable) {
             is PlayerThrowable -> { _errorMessage.value = throwable.message }
             else -> {}
+        }
+    }
+
+    companion object {
+        val Factory = RankFactory(DefaultRankRepository())
+
+        class RankFactory(private val rankRepository: RankRepository) :
+            ViewModelProvider.Factory {
+            override fun <T : ViewModel> create(modelClass: Class<T>): T {
+                return RankViewModel(rankRepository) as T
+            }
         }
     }
 }
